+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Podemos ter pequenas breaking changes sem alterar o `major` version, apesar de s
 Devemos adicionar o comentário `<!-- N/A -->` (Não adicionar), para que não precise adicionar um item do changelog ao lançar uma nova versão stable.
 Caso adicionado no escopo inicial, todos os conteúdos abaixo não serão adicionados. Caso adicionado na linha, será considerado apenas ela.
 
+## Não publicado
+### Adicionado
+- `QasToggle`: Adicionado prop `title` para ter a possibilidade de ter um título acima do toggle.
+
 ## [3.19.0-beta.9] - 10-10-2025
 ### Modificado
 - `QasInput`: Modificado máscara de CNPJ para aceitar letras conforme o novo padrão.